requirements.txt
django
djangorestframework
django-rest-auth
django-allauth

django-jsonfield
django-shortuuidfield
phonenumbers
django-phonenumber-field
django-location-field

geoip2
country_currencies

pillow
django-qsessions
django-money

django-extensions
werkzeug
pyOpenSSL

pyuploadcare
django-sendgrid-v5


google-api-python-client
oauth2client

django-environ
django-split-settings

gunicorn
django_heroku